Reduce size of decode immediate from uint64_t to int32_t
- Add ptr notation to meta/opcode-pseudocode-c and gen-interp
- Update constraints to use integers
- Update generated interpreter using cast to host width

+++ src/asm/riscv-codec.h
@@ -73,7 +73,7 @@ namespace riscv
 
 	struct decode
 	{
-		uint64_t  imm;        /* : 20 (44 variable redundant bits) pre-shifted */
+		int32_t   imm;        /* : 20 (44 variable redundant bits) pre-shifted */
 		uint32_t  rd    : 8;  /* : 5  (3 redundant bits) sized for rw perf */
 		uint32_t  rs1   : 8;  /* : 5  (3 redundant bits) sized for rw perf */
 		uint32_t  rs2   : 8;  /* : 5  (3 redundant bits) sized for rw perf */
